% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/observation_parameters.R
\docType{class}
\name{observation_parameters-class}
\alias{observation_parameters-class}
\alias{response_parameters,observation_parameters-method}
\alias{response_parameters<-,observation_parameters-method}
\alias{fixed_effects,observation_parameters-method}
\alias{fixed_effects<-,observation_parameters-method}
\title{An S4 class to hold observation parameters for a stass model.}
\usage{
\S4method{response_parameters}{observation_parameters}(x)

\S4method{response_parameters}{observation_parameters}(x) <- value

\S4method{fixed_effects}{observation_parameters}(x)

\S4method{fixed_effects}{observation_parameters}(x) <- value
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}
}
\description{
An S4 class to hold observation parameters for a stass model.
}
\section{Functions}{
\itemize{
\item \code{response_parameters(observation_parameters)}: Get response distribution parameters

\item \code{response_parameters(observation_parameters) <- value}: Set response distribution parameters

\item \code{fixed_effects(observation_parameters)}: Get fixed effect parameters

\item \code{fixed_effects(observation_parameters) <- value}: Set fixed effect parameters

}}
\section{Slots}{

\describe{
\item{\code{response_parameters}}{A list containing a data.frame of response
distribution parameters.}

\item{\code{fixed_effects}}{A list containing a data.frame of fixed effect parameters
for each spatio-temporally varying parameter.}
}}

\seealso{
Other stass_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{stass-class}},
\code{\link{tracing-class}}
}
\concept{stass_classes}
\keyword{internal}
