% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/stass.R, R/show.R
\docType{class}
\name{stass-class}
\alias{stass-class}
\alias{settings,stass-method}
\alias{tracing,stass-method}
\alias{time_effects,stass-method}
\alias{time_effects<-,stass-method}
\alias{pg_re,stass-method}
\alias{pg_re<-,stass-method}
\alias{tg_re,stass-method}
\alias{tg_re<-,stass-method}
\alias{random_effects,stass-method}
\alias{persistent_graph,stass-method}
\alias{transient_graph,stass-method}
\alias{graph,stass-method}
\alias{covariance_function,stass-method}
\alias{covariance_function<-,stass-method}
\alias{space_parameters,stass-method}
\alias{space_parameters<-,stass-method}
\alias{time_parameters,stass-method}
\alias{time_parameters<-,stass-method}
\alias{dat,stass-method}
\alias{dat<-,stass-method}
\alias{data_predictions,stass-method}
\alias{data_predictions<-,stass-method}
\alias{response_parameters,stass-method}
\alias{response_parameters<-,stass-method}
\alias{fixed_effects,stass-method}
\alias{fixed_effects<-,stass-method}
\alias{distance_units,stass-method}
\alias{distance_units<-,stass-method}
\alias{formula,stass-method}
\alias{formula<-,stass-method}
\alias{parameters,stass-method}
\alias{parameters<-,stass,parameters-method}
\alias{convergence,stass-method}
\alias{timing,stass-method}
\alias{show,stass-method}
\title{An S4 class describing a stass model.}
\usage{
\S4method{settings}{stass}(x)

\S4method{tracing}{stass}(x)

\S4method{time_effects}{stass}(x)

\S4method{time_effects}{stass}(x) <- value

\S4method{pg_re}{stass}(x)

\S4method{pg_re}{stass}(x) <- value

\S4method{tg_re}{stass}(x)

\S4method{tg_re}{stass}(x) <- value

\S4method{random_effects}{stass}(x)

\S4method{persistent_graph}{stass}(x)

\S4method{transient_graph}{stass}(x)

\S4method{graph}{stass}(x)

\S4method{covariance_function}{stass}(x)

\S4method{covariance_function}{stass}(x) <- value

\S4method{space_parameters}{stass}(x)

\S4method{space_parameters}{stass}(x) <- value

\S4method{time_parameters}{stass}(x)

\S4method{time_parameters}{stass}(x) <- value

\S4method{dat}{stass}(x)

\S4method{dat}{stass}(x) <- value

\S4method{data_predictions}{stass}(x)

\S4method{data_predictions}{stass}(x) <- value

\S4method{response_parameters}{stass}(x)

\S4method{response_parameters}{stass}(x) <- value

\S4method{fixed_effects}{stass}(x)

\S4method{fixed_effects}{stass}(x) <- value

\S4method{distance_units}{stass}(x)

\S4method{distance_units}{stass}(x) <- value

\S4method{formula}{stass}(x)

\S4method{formula}{stass}(x) <- value

\S4method{parameters}{stass}(x)

\S4method{parameters}{stass,parameters}(x) <- value

\S4method{convergence}{stass}(x)

\S4method{timing}{stass}(x)

\S4method{show}{stass}(object)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}

\item{object}{An object}
}
\description{
An S4 class describing a stass model.
}
\section{Functions}{
\itemize{
\item \code{settings(stass)}: Get model settings

\item \code{tracing(stass)}: Get tracing information, see \link{tracing}.

\item \code{time_effects(stass)}: Get temporal random effects

\item \code{time_effects(stass) <- value}: Set temporal random effects

\item \code{pg_re(stass)}: Get persistent graph random effects

\item \code{pg_re(stass) <- value}: Set spatio-temporal random effects

\item \code{tg_re(stass)}: Get transient graph random effects

\item \code{tg_re(stass) <- value}: Set transient graph random effects

\item \code{random_effects(stass)}: Get list containing the persistent graph random
effects and the transient graph random effects.

\item \code{persistent_graph(stass)}: Get persistent graph

\item \code{transient_graph(stass)}: Get transient graph

\item \code{graph(stass)}: Get a list containing the persistent and transient graphs

\item \code{covariance_function(stass)}: Get covariance function

\item \code{covariance_function(stass) <- value}: Set covariance function(s). Run
get_stass_distributions("covariance") for valid covariance functions.
Setting the covariance function also overwrites the spatial parameters.

\item \code{space_parameters(stass)}: Get spatial parameters

\item \code{space_parameters(stass) <- value}: Set spatial parameters

\item \code{time_parameters(stass)}: Get time parameters

\item \code{time_parameters(stass) <- value}: Set time parameters

\item \code{dat(stass)}: Get data including response variables, time indices,
locations, covariates, etc.

\item \code{dat(stass) <- value}: Set data. Warning: if you want to add new
rows you must give each additional row a value in the graph_idx column,
or create a new object from scratch.

\item \code{data_predictions(stass)}: Get data predictions, a long_stars object
with the data (see \code{dat}) and associated random effect
predictions.

\item \code{data_predictions(stass) <- value}: Set data predictions

\item \code{response_parameters(stass)}: Get response distribution parameters

\item \code{response_parameters(stass) <- value}: Set response distribution parameters

\item \code{fixed_effects(stass)}: Get fixed effect parameters

\item \code{fixed_effects(stass) <- value}: Set fixed effect parameters

\item \code{distance_units(stass)}: Get distance units used for the model

\item \code{distance_units(stass) <- value}: Set distance units used for the model

\item \code{formula(stass)}: Get formula used for model

\item \code{formula(stass) <- value}: Set formula used for the model

\item \code{parameters(stass)}: Get all model parameters as a parameters object

\item \code{parameters(x = stass) <- value}: Set all model parameters with a new
parameters object

\item \code{convergence(stass)}: Get convergence message from optimizer

\item \code{timing(stass)}: Get all timing information, see \link{tracing}

\item \code{show(stass)}: Print method

}}
\section{Slots}{

\describe{
\item{\code{process}}{A process object.}

\item{\code{observations}}{An observations object.}

\item{\code{settings}}{A settings object.}
}}

\seealso{
stass_prepare

Other stass_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{tracing-class}}
}
\concept{stass_classes}
